commits
add dproxy ssh-identity checker
add git svn wrap

+++ dproxy.sh
@@ -9,12 +9,14 @@ test_domain=
 
 retry=0
 autoagent=0
+identity=
 
 shift; shift
 while test $# -gt -0; do case $1 in
     --retry ) shift; retry=$1;;
     --test-domain ) shift; test_domain=$1;;
     --ssh-agent ) autoagent=1;;
+    --ssh-identity ) shift; identity=$1;;
     - ) shift; testing=$@; break;;
 esac; shift; done
 testing=$@
@@ -22,7 +24,7 @@ if [ "$testing" = "" ]; then
     google=www.google.com
     [[ "$remote" =~ (bj|cn|ali) ]] && google=www.baidu.com
     [ "$test_domain" != "" ] && google=$test_domain
-    testing="curl -s --socks5-hostname localhost:$port -k -I -o /dev/null -w \"%{http_code}\" https://$google"
+    testing="curl -s --socks5-hostname localhost:$port -k -I -o /dev/null -w %{http_code} https://$google"
 fi
 # >&2 echo "testing = $testing"
 
@@ -78,7 +80,7 @@ if [ "$testing" != "" ]; then
     while test $retry -gt 0; do
         err_resp="`$testing 2>&1 >$ret_file`"
 
-        if [ "$err_resp" != "" ] ; then
+        if [ "$err_resp" != "" ] && [ "$err_resp" != "200" ]; then
             echo "$err_resp"
             rm -f $ret_file
             [[ "$err_resp" =~ $rstopretry ]] && on_error
@@ -98,8 +100,8 @@ if [ "$testing" != "" ]; then
 fi
 
 if [ "$autoagent" -eq 1 ]; then
-    echo -n "verify ssh identities... "
-    identities=`ssh $remote ssh-add -l 2>/dev/null | grep id_rsa | wc -l | awk '{print $1}'`
+    echo -n "verify ssh identities... ${identity:0:6}"
+    identities=`ssh $remote ssh-add -l 2>/dev/null | grep "$identity" | grep id_rsa | wc -l | awk '{print $1}'`
     if [ $identities -eq 0 ]; then
         echo -n "need ssh-add! "
         ssh $remote ssh-add || on_error
